api_rest/requirements.txt
django==1.6.3
djangorestframework
South==0.8.4

